Added App , SplashActivity classes and moved to separate packages
