+++ app/src/main/java/com/rustam/cocktail/recipes/app/App.kt
@@ -1,0 +1,11 @@
+package com.rustam.cocktail.recipes.app
+
+import android.app.Application
+
+class App : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+
+    }
+}
