+++ app/src/main/java/com/rustam/cocktail/recipes/presentation/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.rustam.cocktail.recipes
+package com.rustam.cocktail.recipes.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.rustam.cocktail.recipes.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
