+++ app/src/main/java/com/rustam/cocktail/recipes/presentation/SplashActivity.kt
@@ -1,0 +1,30 @@
+package com.rustam.cocktail.recipes.presentation
+
+import android.content.Intent
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import androidx.appcompat.app.AppCompatActivity
+
+class SplashActivity : AppCompatActivity() {
+
+    private val runnable = Runnable {
+        startActivity(Intent(this, MainActivity::class.java))
+        finishAffinity()
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        Handler(Looper.getMainLooper()).postDelayed(runnable, SPLASH_SHOW_TIME)
+    }
+
+    override fun onPause() {
+        super.onPause()
+        Handler(Looper.getMainLooper()).removeCallbacks(runnable)
+    }
+
+
+    companion object {
+        private const val SPLASH_SHOW_TIME = 400L
+    }
+}
